requirement.txt
django==4.0.3
geopandas==0.10.2
django-debug-toolbar==3.2.4
django-cors-headers==3.11.0
django-crispy-forms==1.14.0
psycopg2==2.9.3
openpyxl==3.0.10